# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Two namespaces are documented: the original `/api/v1` surface and the
    consistent `/api/v2` surface introduced in issue #144. v2 is additive; v1 is
    unchanged. See docs/api-v2.md for the field-by-field v1 -> v2 mapping.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /api/v1/stats/global:
    get:
      tags: [Stats]
      summary: Global index statistics
      operationId: getGlobalStats
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Global counts across all tracked contracts
          content:
            application/json:
              schema:
                type: object
                required: [tracked_contracts, total_events, total_invocations, total_storage_entries]
                properties:
                  tracked_contracts:
                    type: integer
                    format: int64
                  total_events:
                    type: integer
                    format: int64
                  total_invocations:
                    type: integer
                    format: int64
                  total_storage_entries:
                    type: integer
                    format: int64
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts:
    get:
      tags: [Contracts]
      summary: List tracked contracts
      operationId: listContracts
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by contract status (pending, active, backfilling, error, paused).
        - $ref: "#/components/parameters/TagParam"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [id, label, network, status, added_at]
          description: Column to sort by. Defaults to id when omitted; values outside the enum are rejected with 422.
        - name: dir
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction. Defaults to asc when omitted; values outside the enum are rejected with 422.
      responses:
        "200":
          description: Page of contracts
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contract"
                  next_cursor:
                    type: string
                    description: Opaque cursor for the next page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Register a contract for tracking
      operationId: registerContract
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, network]
              properties:
                id:
                  type: string
                  pattern: "^C[A-Za-z0-9]{55}$"
                  description: Stellar contract ID (56 chars, starts with C).
                network:
                  type: string
                  enum: [testnet, mainnet, futurenet, standalone]
                label:
                  type: string
                  description: Optional human-friendly label.
      responses:
        "201":
          description: Contract registered (status pending until backfill completes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/batch:
    post:
      tags: [Contracts]
      summary: Apply a bulk action to many contracts
      operationId: batchContracts
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchContractsRequest"
      responses:
        "200":
          description: Bulk action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchContractsResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/tags:
    post:
      tags: [Contracts]
      summary: Add a tag to a contract
      description: |
        Adds a user-defined tag to a contract. The tag is trimmed and
        lowercased before storage. Adding a tag that is already present is a
        no-op, so the operation is idempotent.
      operationId: addContractTag
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractTagRequest"
      responses:
        "200":
          description: The contract's full, sorted tag list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractTags"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/tags/{tag}:
    delete:
      tags: [Contracts]
      summary: Remove a tag from a contract
      description: |
        Removes a user-defined tag from a contract. Removing a tag that is not
        present is a no-op.
      operationId: removeContractTag
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/TagPath"
      responses:
        "204":
          description: Tag removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/events:
    get:
      tags: [Events]
      summary: List contract events
      operationId: listContractEvents
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Event type filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
      responses:
        "200":
          description: Page of events
          content:
            application/json:
              schema:
                type: object
                required: [events, next_cursor]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/events.csv:
    get:
      tags: [Events]
      summary: Export contract events as CSV
      description: >
        Streams every matching event as CSV instead of paginating them, so a
        large contract can be handed to a spreadsheet or an analyst in one
        request. Rows are ordered by ledger ascending, then event id, and the
        file is deterministic for an unchanged store.

        Unlike the JSON listing there is no cursor and no row cap: an export is
        meant to be complete. The same `type`, `from` and `to` filters apply,
        but `topic` and `in_successful_call` are not supported here.

        The header row is always present, so an unknown contract yields a
        file with no data rows rather than a 404. Free-text columns are
        prefixed with an apostrophe when they start with a character a
        spreadsheet would treat as a formula.
      operationId: exportContractEventsCSV
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/NetworkParam"
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Event type filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive). Must not exceed `to`.
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive). Must not be below `from`.
      responses:
        "200":
          description: >
            CSV export of matching contract events, served as a downloadable
            attachment with `Content-Type: text/csv; charset=utf-8` and
            `Content-Disposition: attachment; filename="<id>-events.csv"`.
            The columns are id, contract_id, network, ledger,
            ledger_closed_at, tx_hash, type, topic_xdr, value_xdr,
            topic_decoded, value_decoded, in_successful_call.
          content:
            text/csv:
              schema:
                type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/invocations:
    get:
      tags: [Invocations]
      summary: List contract invocations
      operationId: listContractInvocations
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [SUCCESS, FAILED, NOT_FOUND]
          description: Invocation status filter.
        - name: function_name
          in: query
          required: false
          schema:
            type: string
          description: Function name filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
      responses:
        "200":
          description: Page of invocations
          content:
            application/json:
              schema:
                type: object
                required: [invocations, next_cursor]
                properties:
                  invocations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invocation"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/storage:
    get:
      tags: [Storage]
      summary: List contract storage entries
      operationId: listContractStorage
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: durability
          in: query
          required: false
          schema:
            type: string
            enum: [temporary, persistent, instance]
          description: Storage durability filter.
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [live, archived, deleted]
          description: Storage entry status filter.
      responses:
        "200":
          description: Page of storage entries
          content:
            application/json:
              schema:
                type: object
                required: [storage, next_cursor]
                properties:
                  storage:
                    type: array
                    items:
                      $ref: "#/components/schemas/StorageEntry"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/stats:
    get:
      tags: [Stats]
      summary: Per-contract statistics
      operationId: getContractStats
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: window
          in: query
          required: false
          schema:
            type: string
          description: Aggregation window. 7d and 30d map to 7/30 days; anything else is 24 hours.
      responses:
        "200":
          description: Contract statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractStats"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/forecast:
    get:
      tags: [Forecast]
      summary: Forecast contract usage over the next N days
      operationId: getContractForecast
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: horizon
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
          description: Number of days to forecast (default 30; optional trailing d accepted).
      responses:
        "200":
          description: Forecast series for fees, invocations, and events
          content:
            application/json:
              schema:
                type: object
                required: [contract_id, lookback_days, series]
                properties:
                  contract_id:
                    type: string
                  lookback_days:
                    type: integer
                  series:
                    type: array
                    items:
                      $ref: "#/components/schemas/ForecastSeries"
        "400":
          description: Invalid horizon parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/snapshot:
    get:
      tags: [Snapshots]
      summary: Snapshot of contract storage at a historical ledger
      operationId: getContractSnapshot
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: ledger
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Ledger to snap (greater than zero).
      responses:
        "200":
          description: Storage snapshot plus the last event at or before the ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/verify:
    post:
      tags: [Verification]
      summary: Verify contract source against the on-chain Wasm hash
      operationId: verifyContractSource
      description: |
        Rebuilds the submitted source in an isolated sandbox and compares the
        resulting SHA-256 Wasm hash with the hash recorded on-chain for this
        contract. Supply exactly one source payload: `source.kind` is either
        `archive` (with `source.content_base64`) or `git` (with `source.url`
        and `source.commit`). `source.subdir` and `artifact` are optional
        overrides for multi-crate sources and for naming the built `.wasm`
        directly.

        A build that runs but does not match is a successful request: it
        returns 200 with `status: failed`, `matched: false`, and diagnostics
        explaining the likely cause. Use 422 for malformed input: an unknown
        `kind`, a missing payload, or a body over 16 MiB. A deployment without
        a build sandbox answers 503 rather than pretending to verify.
      security:
        - ApiKeyAuth: [write:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractVerificationRequest"
      responses:
        "200":
          description: Verification verdict, including a hash mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractVerification"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: No build sandbox is configured on this deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/contracts/{id}/verification:
    get:
      tags: [Verification]
      summary: Read the stored source verification for a contract
      operationId: getContractVerification
      description: |
        Returns the most recent verification record for the contract. Responds
        404 both when the contract itself is unknown and when the contract
        exists but has never been submitted for source verification.
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: Stored verification record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractVerification"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/stream:
    get:
      tags: [Events]
      summary: Poll the most recent contract events
      operationId: streamContractEvents
      description: |
        Short-poll convenience used by the dashboard. Returns the 20 most
        recent events as JSON (not SSE).
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: Most recent events
          headers:
            Last-Modified:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke an API key
      operationId: revokeApiKey
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID.
      responses:
        "204":
          description: Key revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/watchdog/stats:
    get:
      tags: [Watchdog]
      summary: Watchdog statistics
      operationId: getWatchdogStats
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/NetworkParam"
      responses:
        "200":
          description: Watchdog summary counts
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchdogStats"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/alerts:
    get:
      tags: [Watchdog]
      summary: List watchdog alerts
      operationId: listWatchdogAlerts
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [Info, Warning, Critical]
          description: Severity filter.
        - $ref: "#/components/parameters/NetworkParam"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
          description: Maximum alerts to return (default 100).
        - $ref: "#/components/parameters/AlertCursor"
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                required: [alerts, next_cursor]
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContractAlert"
                  next_cursor:
                    type: string
                    description: Cursor for the next (older) page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Global per-IP rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...